DEV: Migrate insert-hyperlink to the new modal api
