+++ app/assets/javascripts/discourse/app/components/modal/insert-hyperlink.hbs
@@ -1,0 +1,74 @@
+{{! template-lint-disable no-pointer-down-event-binding }}
+<DModal
+  {{on "keydown" this.keyDown}}
+  {{on "mousedown" this.mouseDown}}
+  @closeModal={{@closeModal}}
+  @title={{i18n "composer.link_dialog_title"}}
+  @bodyClass="insert-link"
+  class="insert-hyperlink-modal"
+>
+  <:body>
+    <div class="inputs">
+      <input
+        {{on "input" this.search}}
+        value={{this.linkUrl}}
+        placeholder={{i18n "composer.link_url_placeholder"}}
+        type="text"
+        autofocus="autofocus"
+        class="link-url"
+      />
+
+      {{#if this.searchLoading}}
+        {{loading-spinner}}
+      {{/if}}
+
+      {{#if this.searchResults}}
+        <div class="internal-link-results">
+          {{#each this.searchResults as |result|}}
+            <a
+              {{on "click" this.linkClick}}
+              href={{result.url}}
+              data-title={{result.fancy_title}}
+              class="search-link"
+            >
+              <TopicStatus @topic={{result}} @disableActions={{true}} />
+              {{replace-emoji result.title}}
+              <div class="search-category">
+                {{#if result.category.parentCategory}}
+                  {{category-link result.category.parentCategory}}
+                {{/if}}
+                {{category-link result.category hideParent=true}}
+                {{discourse-tags result}}
+              </div>
+            </a>
+          {{/each}}
+        </div>
+      {{/if}}
+    </div>
+
+    <div class="inputs">
+      <input
+        {{on "input" this.updateLinkText}}
+        value={{this.linkText}}
+        placeholder={{i18n "composer.link_optional_text"}}
+        type="text"
+        class="link-text"
+      />
+    </div>
+  </:body>
+
+  <:footer>
+    <DButton
+      @action={{this.ok}}
+      @label="composer.modal_ok"
+      type="submit"
+      class="btn-primary"
+    />
+
+    <DButton
+      @action={{@closeModal}}
+      @label="composer.modal_cancel"
+      class="btn-danger"
+    />
+  </:footer>
+</DModal>
