+++ app/assets/javascripts/discourse/app/services/modal.js
@@ -33,7 +33,6 @@ const KNOWN_LEGACY_MODALS = [
   "history",
   "ignore-duration-with-username",
   "ignore-duration",
-  "insert-hyperlink",
   "jump-to-post",
   "login",
   "move-to-topic",
